steam-library-manager/requirements.txt
psutil==5.9.6
python-dotenv==1.0.1
tqdm==4.66.1
vdf==3.4